refactor(agent): extract atomic text write helper (#212)
