+++ crates/pi-coding-agent/src/main.rs
@@ -1,3 +1,4 @@
+mod atomic_io;
 mod auth_commands;
 mod bootstrap_helpers;
 mod channel_store;
@@ -53,6 +54,7 @@ use pi_ai::{
 use serde::{Deserialize, Serialize};
 use serde_json::Value;
 
+pub(crate) use crate::atomic_io::write_text_atomic;
 pub(crate) use crate::auth_commands::execute_auth_command;
 #[cfg(test)]
 pub(crate) use crate::auth_commands::{parse_auth_command, AuthCommand};
@@ -1844,42 +1846,6 @@ async fn main() -> Result<()> {
         return Ok(());
     }
     run_interactive(agent, session_runtime, interactive_config).await
-}
-
-pub(crate) fn write_text_atomic(path: &Path, content: &str) -> Result<()> {
-    if path.as_os_str().is_empty() {
-        bail!("destination path cannot be empty");
-    }
-    if path.exists() && path.is_dir() {
-        bail!("destination path '{}' is a directory", path.display());
-    }
-
-    let parent_dir = path
-        .parent()
-        .filter(|dir| !dir.as_os_str().is_empty())
-        .unwrap_or_else(|| Path::new("."));
-    std::fs::create_dir_all(parent_dir)
-        .with_context(|| format!("failed to create {}", parent_dir.display()))?;
-
-    let temp_name = format!(
-        ".{}.tmp-{}-{}",
-        path.file_name()
-            .and_then(|name| name.to_str())
-            .unwrap_or("session-graph"),
-        std::process::id(),
-        current_unix_timestamp()
-    );
-    let temp_path = parent_dir.join(temp_name);
-    std::fs::write(&temp_path, content)
-        .with_context(|| format!("failed to write temporary file {}", temp_path.display()))?;
-    std::fs::rename(&temp_path, path).with_context(|| {
-        format!(
-            "failed to rename temporary graph file {} to {}",
-            temp_path.display(),
-            path.display()
-        )
-    })?;
-    Ok(())
 }
 
 #[cfg(test)]
